refactor: split tui activity helpers
